Featured Image: Fix overlay rendering in the editor (#60187)
